+++ assets/js/script.js
@@ -1,0 +1,1 @@
+/*----- custom addon js -----*/